fix: 수익률 계산 오류 수정

+++ src/main/java/lotto/domain/User.java
@@ -39,6 +39,6 @@ public class User {
     }
 
     public double getYield() {
-        return benefit / MONEY * 100;
+        return (double) benefit / MONEY * 100;
     }
 }
